requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.14
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.9
zipp==3.8.0
